AI: Merge app threads to hybrid.
