+++ trunk/src/app/srs_app_log.hpp
@@ -15,8 +15,6 @@
 #include <srs_app_reload.hpp>
 #include <srs_protocol_log.hpp>
 
-class SrsThreadMutex;
-
 // For log TAGs.
 #define TAG_MAIN "MAIN"
 #define TAG_MAYBE "MAYBE"
@@ -42,9 +40,6 @@ private:
     bool log_to_file_tank;
     // Whether use utc time.
     bool utc;
-    // TODO: FIXME: use macro define like SRS_MULTI_THREAD_LOG to switch enable log mutex or not.
-    // Mutex for multithread log.
-    SrsThreadMutex *mutex_;
 
 public:
     SrsFileLog();
