+++ trunk/src/app/srs_app_rtc_conn.cpp
@@ -33,7 +33,6 @@ using namespace std;
 #include <srs_app_source.hpp>
 #include <srs_app_srt_source.hpp>
 #include <srs_app_statistic.hpp>
-#include <srs_app_threads.hpp>
 #include <srs_app_utility.hpp>
 #include <srs_core_autofree.hpp>
 #include <srs_kernel_buffer.hpp>
