+++ trunk/src/app/srs_app_rtc_queue.cpp
@@ -13,7 +13,6 @@
 using namespace std;
 
 #include <srs_app_circuit_breaker.hpp>
-#include <srs_app_threads.hpp>
 #include <srs_app_utility.hpp>
 #include <srs_kernel_error.hpp>
 #include <srs_kernel_rtc_rtp.hpp>
